requirements.txt
streamlit
langchain>=0.2.7
langchain-core>=0.2.5
langchain-openai>=0.1.7
langchain-pinecone>=0.1.2
langchain-community>=0.2.6
pinecone-client>=5.0.1
openai>=1.45.0
tiktoken
python-dotenv
pdfplumber
docx2txt
pytesseract
Pillow
pdf2image
docling
langsmith[evaluation]>=0.1.37
cohere
sentence-transformers
torch